refactor: Reorganise core into registry, types, constructors, assertions
