+++ internal/core/assertions.go
